// src/app/cts/shared/models/timetable.ts
export class Timetable {

         id: number;
         classid: number;
         subjectid: number;
         teacherid: number;
         periodfrom: Date;
         periodto: Date;
         createddate: Date;
         createdby: string;
         indexId:number;
        querytype:number;
        status:string;
    
}